Simplify image worker scheduling
